fixed furnace gui not showing burn for lava buckets
(cherry picked from commit fc9171f2285419dd16678752fdb4ccdae01439e0)

+++ src/main/java/mod/alexndr/machines/api/content/AbstractModFurnaceContainer.java
@@ -38,8 +38,8 @@ public abstract class AbstractModFurnaceContainer<T extends AbstractModFurnaceBl
         // Add tracking for data (Syncs to client/updates value when it changes)
         this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.smeltTimeLeft, v -> tileEntity.smeltTimeLeft = (short) v));
         this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.maxSmeltTime, v -> tileEntity.maxSmeltTime = (short) v));
-        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.fuelBurnTimeLeft, v -> tileEntity.fuelBurnTimeLeft = (short) v));
-        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.maxFuelBurnTime, v -> tileEntity.maxFuelBurnTime = (short) v));
+        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.fuelBurnTimeLeft, v -> tileEntity.fuelBurnTimeLeft = (int) v));
+        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.maxFuelBurnTime, v -> tileEntity.maxFuelBurnTime = (int) v));
 
         // Tile inventory slot(s)
         this.addSlot(new SlotItemHandler(tileEntity.inventory, AbstractModFurnaceTileEntity.FUEL_SLOT, 56, 53));
